inline getadd stuff onto the trait
I left some support code in `getadd`.

I also went and changed the input arguments to take &Path instead of PathBuf
