+++ iroh/src/getadd.rs
@@ -1,93 +1,56 @@
 use std::path::{Path, PathBuf};
 
+use crate::api::{Iroh, OutType};
 use anyhow::{Context, Result};
-use cid::Cid;
 use futures::Stream;
 use futures::StreamExt;
 use iroh_resolver::resolver::Path as IpfsPath;
-use iroh_resolver::{resolver, unixfs_builder};
 use iroh_rpc_client::Client;
 
-pub async fn get(client: &Client, ipfs_path: &IpfsPath, output: Option<&Path>) -> Result<()> {
-    let blocks = get_stream(client.clone(), ipfs_path, output.map(|p| p.to_path_buf()));
-    tokio::pin!(blocks);
-    while let Some(block) = blocks.next().await {
-        let (path, out) = block?;
-        match out {
-            OutType::Dir => {
-                tokio::fs::create_dir_all(path).await?;
-            }
-            OutType::Reader(mut reader) => {
-                if let Some(parent) = path.parent() {
-                    tokio::fs::create_dir_all(parent).await?;
+impl<'a> Iroh<'a> {
+    pub fn get_stream<'b>(
+        &self,
+        root: &'b IpfsPath,
+        output: Option<&'b Path>,
+    ) -> impl Stream<Item = Result<(PathBuf, OutType<Client>)>> + 'b {
+        tracing::debug!("get {:?}", root);
+        let resolver = iroh_resolver::resolver::Resolver::new(self.get_client().clone());
+        let results = resolver.resolve_recursive_with_paths(root.clone());
+        async_stream::try_stream! {
+            tokio::pin!(results);
+            while let Some(res) = results.next().await {
+                let (path, out) = res?;
+                let path = Self::make_output_path(path, root.clone(), output.clone())?;
+                if out.is_dir() {
+                    yield (path, OutType::Dir);
+                } else {
+                    let reader = out.pretty(resolver.clone(), Default::default())?;
+                    yield (path, OutType::Reader(reader));
                 }
-                let mut f = tokio::fs::File::create(path).await?;
-                tokio::io::copy(&mut reader, &mut f).await?;
             }
         }
     }
-    Ok(())
-}
 
-pub async fn add(client: &Client, path: &Path, recursive: bool, wrap: bool) -> Result<Cid> {
-    let providing_client = iroh_resolver::unixfs_builder::StoreAndProvideClient {
-        client: Box::new(client),
-    };
-    if path.is_dir() {
-        unixfs_builder::add_dir(Some(&providing_client), path, wrap, recursive).await
-    } else if path.is_file() {
-        unixfs_builder::add_file(Some(&providing_client), path, wrap).await
-    } else {
-        anyhow::bail!("can only add files or directories");
-    }
-}
-
-pub enum OutType<T: resolver::ContentLoader> {
-    Dir,
-    Reader(resolver::OutPrettyReader<T>),
-}
-
-pub fn get_stream<'a>(
-    client: Client,
-    root: &'a IpfsPath,
-    output: Option<PathBuf>,
-) -> impl Stream<Item = Result<(PathBuf, OutType<Client>)>> + 'a {
-    tracing::debug!("get {:?}", root);
-    let resolver = iroh_resolver::resolver::Resolver::new(client);
-    let results = resolver.resolve_recursive_with_paths(root.clone());
-    async_stream::try_stream! {
-        tokio::pin!(results);
-        while let Some(res) = results.next().await {
-            let (path, out) = res?;
-            let path = make_output_path(path, root.clone(), output.clone())?;
-            if out.is_dir() {
-                yield (path, OutType::Dir);
-            } else {
-                let reader = out.pretty(resolver.clone(), Default::default())?;
-                yield (path, OutType::Reader(reader));
-            }
+    /// Adjusts the full path to replace the root with any given output path
+    /// if it exists.
+    fn make_output_path(full: IpfsPath, root: IpfsPath, output: Option<&Path>) -> Result<PathBuf> {
+        if let Some(output) = output {
+            let root_str = &root.to_string()[..];
+            let full_as_path = PathBuf::from(full.to_string());
+            let path_str = full_as_path.to_str().context("invalid root path")?;
+            let output_str = output.to_str().context("invalid output path")?;
+            Ok(PathBuf::from(path_str.replace(root_str, output_str)))
+        } else {
+            // returns path as a string
+            Ok(PathBuf::from(full.to_string_without_type()))
         }
-    }
-}
-
-/// Adjusts the full path to replace the root with any given output path
-// if it exists.
-fn make_output_path(full: IpfsPath, root: IpfsPath, output: Option<PathBuf>) -> Result<PathBuf> {
-    if let Some(ref output) = output {
-        let root_str = &root.to_string()[..];
-        let full_as_path = PathBuf::from(full.to_string());
-        let path_str = full_as_path.to_str().context("invalid root path")?;
-        let output_str = output.to_str().context("invalid output path")?;
-        Ok(PathBuf::from(path_str.replace(root_str, output_str)))
-    } else {
-        // returns path as a string
-        Ok(PathBuf::from(full.to_string_without_type()))
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use iroh_resolver::resolver;
     use std::str::FromStr;
 
     #[test]
@@ -100,7 +63,7 @@ mod tests {
                 .unwrap();
         let output = Some(PathBuf::from("foo"));
         let expect = PathBuf::from("foo/bar.txt");
-        let got = make_output_path(full, root, output).unwrap();
+        let got = Iroh::make_output_path(full, root, output.as_deref()).unwrap();
         assert_eq!(expect, got);
 
         // test with output filepath
@@ -114,7 +77,7 @@ mod tests {
         .unwrap();
         let output = Some(PathBuf::from("foo/baz.txt"));
         let expect = PathBuf::from("foo/baz.txt");
-        let got = make_output_path(full, root, output).unwrap();
+        let got = Iroh::make_output_path(full, root, output.as_deref()).unwrap();
         assert_eq!(expect, got);
 
         // test no output path
@@ -126,7 +89,7 @@ mod tests {
         .unwrap();
         let output = None;
         let expect = PathBuf::from("QmYbcW4tXLXHWw753boCK8Y7uxLu5abXjyYizhLznq9PUR/bar.txt");
-        let got = make_output_path(full, root, output).unwrap();
+        let got = Iroh::make_output_path(full, root, output).unwrap();
         assert_eq!(expect, got);
     }
 }
